(exp) refactored classes into separate files
